requirements.txt
Flask
pandas
pyodbc
openpyxl
Werkzeug
python-dotenv
Flask-WTF
psycopg2-binary